added condition to only update policies if there is any

+++ x-pack/plugins/fleet/server/services/agent_policy.ts
@@ -1011,38 +1011,42 @@ class AgentPolicyService {
       const { policiesWithSingleAP: packagePoliciesToDelete, policiesWithMultipleAP } =
         this.packagePoliciesWithoutMultiplePolicies(packagePolicies);
 
-      await packagePolicyService.delete(
-        soClient,
-        esClient,
-        packagePoliciesToDelete.map((p) => p.id),
-        {
-          force: options?.force,
-          skipUnassignFromAgentPolicies: true,
-        }
-      );
-      logger.debug(
-        `Deleted package policies with ids ${packagePoliciesToDelete
-          .map((policy) => policy.id)
-          .join(', ')}`
-      );
+      if (packagePoliciesToDelete.length > 0) {
+        await packagePolicyService.delete(
+          soClient,
+          esClient,
+          packagePoliciesToDelete.map((p) => p.id),
+          {
+            force: options?.force,
+            skipUnassignFromAgentPolicies: true,
+          }
+        );
+        logger.debug(
+          `Deleted package policies with ids ${packagePoliciesToDelete
+            .map((policy) => policy.id)
+            .join(', ')}`
+        );
+      }
 
-      await packagePolicyService.bulkUpdate(
-        soClient,
-        esClient,
-        policiesWithMultipleAP.map((policy) => {
-          const newPolicyIds = policy.policy_ids.filter((policyId) => policyId !== id);
-          return {
-            ...policy,
-            policy_id: newPolicyIds[0],
-            policy_ids: newPolicyIds,
-          };
-        })
-      );
-      logger.debug(
-        `Updated package policies with ids ${policiesWithMultipleAP
-          .map((policy) => policy.id)
-          .join(', ')}`
-      );
+      if (policiesWithMultipleAP.length > 0) {
+        await packagePolicyService.bulkUpdate(
+          soClient,
+          esClient,
+          policiesWithMultipleAP.map((policy) => {
+            const newPolicyIds = policy.policy_ids.filter((policyId) => policyId !== id);
+            return {
+              ...policy,
+              policy_id: newPolicyIds[0],
+              policy_ids: newPolicyIds,
+            };
+          })
+        );
+        logger.debug(
+          `Updated package policies with ids ${policiesWithMultipleAP
+            .map((policy) => policy.id)
+            .join(', ')}`
+        );
+      }
     }
 
     if (agentPolicy.is_preconfigured && !options?.force) {
